Update homebrew.md
Was able to get Airsonic to successfully launch at  system boot -- requiring NO user login -- by copying the .plist file to /Library/LuanchDaemons/ and chowning to be owned by root/wheel.  It works, but someone may want to add suggestions for how this can be done most securely, and also discuss how this is not a good idea for mobile/laptop devices running on batteries, etc.

Tested this by rebooting and verifying that Airsonic was working.